usbc: change type for functions that may return error code
TCPC API functions should be able to inform the caller if the function
is supported and successfully executed. Bool values doesn't allow this
so it is needed to change the type to int.
For is_rx_pending_msg function the return code should conform to
existing error codes, so in case of function being not supported,
the -ENOSYS should be returned. In case of successful execution,
if there is no pending message, the -ENODATA should be returned and
in case of message pending, the value of 0.

+++ subsys/usb/usb_c/usbc_prl.c
@@ -446,7 +446,7 @@ static void increment_msgid_counter(const struct device *dev)
 /**
  * @brief Get the SOP* header for the current received message
  */
-static uint32_t get_sop_star_header(const struct device *dev)
+static uint16_t get_sop_star_header(const struct device *dev)
 {
 	struct usbc_port_data *data = dev->data;
 	struct protocol_layer_tx_t *prl_tx = data->prl_tx;
